backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
joblib==1.4.2
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
xgboost==2.1.4
supabase==2.28.3
PyJWT==2.12.1
python-dotenv==1.1.0
pytest==8.3.5
statsmodels==0.14.2
pytest-cov==6.0.0